requirements.txt
streamlit
yfinance
pandas
numpy
scipy
plotly
google-generativeai
